Context switching in asm. I feel much better now.
I think the general problem is this: we can't do the context switch
within a Rust stack frame, because Rust's machine model assumes
that within a stack frame, the stack doesn't get ripped away from
under its fingers.

This is of course usually a reasonable assumption, and so Rust puts
code like destructors "between the last line of code of the function
and the closing brace".

Of course, this breaks if you take away the stack using inline assembly.
I'd assume this is one of the reasons why people don't write Kernel
modules in C++.

So, actually, our context switching logic isn't half bad. It just doesn't
aligh with Rust's memory model. However, it seems like it's OK to just
do the context switching from within a stack frame deeper (that's
guaranteed to be non-Rust, i.e. asm).

The method used for passing parameters here (static variables) is of
course super fucking ugly. It's the easiest way I could think of for
porting, though. Actually, I think we could just to the call to
context_switch() directly inside of reschedule(). I don't think this
conflicts with the above safety precautions, except maybe for the
case where a task context-switches away forever, i.e. when exiting...
or maybe this isn't a problem?

Wait, actually, I think this is *always* a problem when cleaning up;
the same goes for, like, stuff allocated within the outermost entry
point of the task. It seems like the actually correct way would be to
unwind the stack here, calling destructors as we go. Of course, then
we are doing stack unwinding. I'm not sure we can get around that,
though...